lighting/requirements.txt
# Core dependencies
python-osc>=1.9.0
PyYAML>=6.0

# Backend libraries (install only what you need)
python-kasa>=0.6.0        # For Kasa backend
wyze-sdk>=1.3.0           # For Wyze backend
requests>=2.31.0          # For WLED backend (HTTP check)
